requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
graphviz==0.20.1
gurobipy==10.0.2
idna==3.4
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
littleutils==0.2.2
llvmlite==0.40.1rc1
matplotlib==3.7.1
networkx==3.0
numba==0.57.0
numpy==1.24.2
outdated==0.2.2
packaging==23.1
pandas==1.5.3
pandas-flavor==0.5.0
patsy==0.5.3
Pillow==9.5.0
pingouin==0.5.3
pydot==1.4.2
pygobnilp==1.0
pygraphviz==1.11
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==19.0.2
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
statsmodels==0.14.0
tabulate==0.9.0
threadpoolctl==3.1.0
tqdm==4.65.0
urllib3==2.0.3
xarray==2023.5.0
zipp==3.15.0
